chore: restructure ai agent service and decouple it from ai service
