+++ packages/backend/src/ee/clients/OpenAi.ts
@@ -7,11 +7,7 @@ import { StringOutputParser } from '@langchain/core/output_parsers';
 import { LLMResult } from '@langchain/core/outputs';
 import { ChatPromptTemplate } from '@langchain/core/prompts';
 import { Runnable } from '@langchain/core/runnables';
-import {
-    ChatOpenAI,
-    ChatOpenAICallOptions,
-    OpenAIEmbeddings,
-} from '@langchain/openai';
+import { ChatOpenAI, ChatOpenAICallOptions } from '@langchain/openai';
 import { sleep, UnexpectedServerError } from '@lightdash/common';
 
 const DEFAULT_RETRY_TIMEOUT_MS = 5000;
@@ -36,8 +32,6 @@ export default class OpenAi {
 
     model: ChatOpenAI<ChatOpenAICallOptions> | undefined;
 
-    embedder: OpenAIEmbeddings | undefined;
-
     constructor() {
         this.openAiApiKey = process.env.OPENAI_API_KEY;
 
@@ -46,15 +40,6 @@ export default class OpenAi {
                   openAIApiKey: this.openAiApiKey,
                   modelName: process.env.OPENAI_MODEL_NAME ?? 'gpt-4o',
                   temperature: 0.2,
-              })
-            : undefined;
-
-        this.embedder = this.openAiApiKey
-            ? new OpenAIEmbeddings({
-                  apiKey: this.openAiApiKey,
-                  modelName:
-                      process.env.OPENAI_EMBEDDING_MODEL_NAME ??
-                      'text-embedding-3-small',
               })
             : undefined;
     }
