+++ packages/backend/src/ee/clients/Slack/SlackBot.ts
@@ -1,4 +1,8 @@
 import {
+    AiAgentNotFoundError,
+    AiDuplicateSlackPromptError,
+} from '@lightdash/common';
+import {
     AllMiddlewareArgs,
     App,
     Block,
@@ -11,19 +15,15 @@ import { AiAgentModel } from '../../models/AiAgentModel';
 import { AiModel } from '../../models/AiModel';
 import { CommercialSlackAuthenticationModel } from '../../models/CommercialSlackAuthenticationModel';
 import { CommercialSchedulerClient } from '../../scheduler/SchedulerClient';
-import { AiService } from '../../services/AiService/AiService';
 import {
     FollowUpTools,
     followUpToolsText,
-} from '../../services/AiService/utils/aiCopilot/followUpTools';
-import {
-    AiAgentNotFoundError,
-    AiDuplicateSlackPromptError,
-} from '../../services/AiService/utils/errors';
+} from '../../services/ai/types/followUpTools';
+import { AiAgentService } from '../../services/AiAgentService';
 
 type CommercialSlackBotArguments = SlackBotArguments & {
     schedulerClient: CommercialSchedulerClient;
-    aiService: AiService;
+    aiAgentService: AiAgentService;
     aiModel: AiModel;
     aiAgentModel: AiAgentModel;
     slackAuthenticationModel: CommercialSlackAuthenticationModel;
@@ -32,7 +32,7 @@ type CommercialSlackBotArguments = SlackBotArguments & {
 export class CommercialSlackBot extends SlackBot {
     private readonly schedulerClient: CommercialSchedulerClient;
 
-    private readonly aiService: AiService;
+    private readonly aiAgentService: AiAgentService;
 
     private readonly aiModel: AiModel;
 
@@ -43,7 +43,7 @@ export class CommercialSlackBot extends SlackBot {
     constructor(args: CommercialSlackBotArguments) {
         super(args);
         this.schedulerClient = args.schedulerClient;
-        this.aiService = args.aiService;
+        this.aiAgentService = args.aiAgentService;
         this.aiModel = args.aiModel;
         this.aiAgentModel = args.aiAgentModel;
         this.slackAuthenticationModel = args.slackAuthenticationModel;
@@ -110,7 +110,7 @@ export class CommercialSlackBot extends SlackBot {
                         if (!promptUuid) {
                             return;
                         }
-                        await this.aiService.updateHumanScoreForPrompt(
+                        await this.aiAgentService.updateHumanScoreForPrompt(
                             promptUuid,
                             1,
                         );
@@ -155,7 +155,7 @@ export class CommercialSlackBot extends SlackBot {
                         if (!promptUuid) {
                             return;
                         }
-                        await this.aiService.updateHumanScoreForPrompt(
+                        await this.aiAgentService.updateHumanScoreForPrompt(
                             promptUuid,
                             -1,
                         );
@@ -231,18 +231,21 @@ export class CommercialSlackBot extends SlackBot {
 
                             try {
                                 [slackPromptUuid] =
-                                    await this.aiService.createSlackPrompt({
-                                        userUuid,
-                                        projectUuid:
-                                            prevSlackPrompt.projectUuid,
-                                        slackUserId: context.botUserId,
-                                        slackChannelId: channel.id,
-                                        slackThreadTs:
-                                            prevSlackPrompt.slackThreadTs,
-                                        prompt: response.message.text,
-                                        promptSlackTs: response.ts,
-                                        agentUuid: prevSlackPrompt.agentUuid,
-                                    });
+                                    await this.aiAgentService.createSlackPrompt(
+                                        {
+                                            userUuid,
+                                            projectUuid:
+                                                prevSlackPrompt.projectUuid,
+                                            slackUserId: context.botUserId,
+                                            slackChannelId: channel.id,
+                                            slackThreadTs:
+                                                prevSlackPrompt.slackThreadTs,
+                                            prompt: response.message.text,
+                                            promptSlackTs: response.ts,
+                                            agentUuid:
+                                                prevSlackPrompt.agentUuid,
+                                        },
+                                    );
                             } catch (e) {
                                 if (e instanceof AiDuplicateSlackPromptError) {
                                     Logger.debug(
@@ -310,7 +313,7 @@ export class CommercialSlackBot extends SlackBot {
             name = agentConfig.name;
 
             [slackPromptUuid, createdThread] =
-                await this.aiService.createSlackPrompt({
+                await this.aiAgentService.createSlackPrompt({
                     userUuid,
                     projectUuid: agentConfig.projectUuid,
                     slackUserId: event.user,
