+++ packages/backend/src/ee/index.ts
@@ -89,35 +89,27 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                 }),
             aiService: ({ repository, context, models, clients }) =>
                 new AiService({
+                    lightdashConfig: context.lightdashConfig,
                     analytics: context.lightdashAnalytics,
                     dashboardModel: models.getDashboardModel(),
                     dashboardSummaryModel: models.getDashboardSummaryModel(),
                     projectService: repository.getProjectService(),
-                    catalogService:
-                        repository.getCatalogService() as CommercialCatalogService,
-                    userModel: models.getUserModel(),
-                    aiModel: models.getAiModel(),
-                    aiAgentModel: models.getAiAgentModel(),
-                    projectModel: models.getProjectModel(),
                     openAi: new OpenAi(), // TODO This should go in client repository as soon as it is available
-                    slackClient: clients.getSlackClient(),
-                    lightdashConfig: context.lightdashConfig,
-                    organizationModel: models.getOrganizationModel(),
-                    featureFlagService: repository.getFeatureFlagService(),
-                    aiAgentService: repository.getAiAgentService(),
                 }),
             aiAgentService: ({ models, repository, clients, context }) =>
                 new AiAgentService({
+                    lightdashConfig: context.lightdashConfig,
                     analytics: context.lightdashAnalytics,
+                    userModel: models.getUserModel(),
                     aiAgentModel: models.getAiAgentModel(),
-                    slackAuthenticationModel:
-                        models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
                     featureFlagService: repository.getFeatureFlagService(),
                     slackClient: clients.getSlackClient(),
                     aiModel: models.getAiModel(),
                     schedulerClient:
                         clients.getSchedulerClient() as CommercialSchedulerClient,
                     projectService: repository.getProjectService(),
+                    catalogService:
+                        repository.getCatalogService() as CommercialCatalogService,
                 }),
             scimService: ({ models, context }) =>
                 new ScimService({
@@ -280,7 +272,7 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                 s3Client: context.clients.getS3Client(),
                 schedulerClient: context.clients.getSchedulerClient(),
                 slackClient: context.clients.getSlackClient(),
-                aiService: context.serviceRepository.getAiService(),
+                aiAgentService: context.serviceRepository.getAiAgentService(),
                 semanticLayerService:
                     context.serviceRepository.getSemanticLayerService(),
                 catalogService: context.serviceRepository.getCatalogService(),
@@ -295,7 +287,7 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                 slackAuthenticationModel:
                     context.models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
                 unfurlService: context.serviceRepository.getUnfurlService(),
-                aiService: context.serviceRepository.getAiService(),
+                aiAgentService: context.serviceRepository.getAiAgentService(),
                 schedulerClient:
                     context.clients.getSchedulerClient() as CommercialSchedulerClient,
                 aiModel: context.models.getAiModel(),
