+++ packages/backend/src/ee/scheduler/SchedulerWorker.ts
@@ -2,25 +2,25 @@ import { EE_SCHEDULER_TASKS, SchedulerJobStatus } from '@lightdash/common';
 import { SchedulerTaskArguments } from '../../scheduler/SchedulerTask';
 import { SchedulerWorker } from '../../scheduler/SchedulerWorker';
 import { TypedEETaskList } from '../../scheduler/types';
-import { AiService } from '../services/AiService/AiService';
+import { AiAgentService } from '../services/AiAgentService';
 
 type CommercialSchedulerWorkerArguments = SchedulerTaskArguments & {
-    aiService: AiService;
+    aiAgentService: AiAgentService;
 };
 
 export class CommercialSchedulerWorker extends SchedulerWorker {
-    protected readonly aiService: AiService;
+    protected readonly aiAgentService: AiAgentService;
 
     constructor(args: CommercialSchedulerWorkerArguments) {
         super(args);
-        this.aiService = args.aiService;
+        this.aiAgentService = args.aiAgentService;
     }
 
     protected getTaskList(): TypedEETaskList {
         return {
             ...super.getTaskList(),
             [EE_SCHEDULER_TASKS.SLACK_AI_PROMPT]: async (payload, _helpers) => {
-                await this.aiService.replyToSlackPrompt(
+                await this.aiAgentService.replyToSlackPrompt(
                     payload.slackPromptUuid,
                 );
             },
@@ -33,11 +33,13 @@ export class CommercialSchedulerWorker extends SchedulerWorker {
                         helpers.job.id,
                         SchedulerJobStatus.STARTED,
                     );
-                    await this.aiService.generateAgentThreadResponse(
-                        payload.agentUuid,
-                        payload.threadUuid,
-                        payload.promptUuid,
-                    );
+
+                    await this.aiAgentService.generateAgentThreadResponse({
+                        agentUuid: payload.agentUuid,
+                        threadUuid: payload.threadUuid,
+                        promptUuid: payload.promptUuid,
+                    });
+
                     await this.schedulerService.setJobStatus(
                         helpers.job.id,
                         SchedulerJobStatus.COMPLETED,
