+++ packages/backend/src/ee/services/AiAgentService.ts
@@ -2,82 +2,121 @@ import { subject } from '@casl/ability';
 import {
     AiAgentThread,
     AiAgentThreadSummary,
+    AiConversation,
+    AiConversationMessage,
+    AiDuplicateSlackPromptError,
     AiMetricQuery,
+    AiWebAppPrompt,
+    AnyType,
     ApiAiAgentThreadMessageViz,
     ApiCreateAiAgent,
     ApiUpdateAiAgent,
     assertUnreachable,
+    CatalogType,
     CommercialFeatureFlags,
     filterExploreByTags,
     FilterSchemaType,
     ForbiddenError,
+    isSlackPrompt,
     LightdashUser,
     NotFoundError,
     QueryExecutionContext,
+    SlackPrompt,
     UnexpectedServerError,
     type SessionUser,
 } from '@lightdash/common';
-import _ from 'lodash';
+import { CoreMessage } from 'ai';
+import _, { pick } from 'lodash';
+import slackifyMarkdown from 'slackify-markdown';
 import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
 import { type SlackClient } from '../../clients/Slack/SlackClient';
+import { LightdashConfig } from '../../config/parseConfig';
+import Logger from '../../logging/logger';
+import { UserModel } from '../../models/UserModel';
 import { FeatureFlagService } from '../../services/FeatureFlag/FeatureFlagService';
 import { ProjectService } from '../../services/ProjectService/ProjectService';
 import { AiAgentModel } from '../models/AiAgentModel';
 import { AiModel } from '../models/AiModel';
-import type { CommercialSlackAuthenticationModel } from '../models/CommercialSlackAuthenticationModel';
 import { CommercialSchedulerClient } from '../scheduler/SchedulerClient';
-import { isCsvFileConfig, renderCsvFile } from './AiService/charts/csvFile';
+import { runAgent } from './ai/agents/agent';
+import { generateEmbeddingsNameAndDescription } from './ai/embeds/embed';
+import { getChatHistoryFromThreadMessages } from './ai/prompts/conversationHistory';
 import {
-    isTimeSeriesMetricChartConfig,
-    renderTimeseriesChart,
-} from './AiService/charts/timeSeriesChart';
-import { AiAgentVizConfig } from './AiService/charts/types';
+    GetExploreFn,
+    GetPromptFn,
+    RunMiniMetricQueryFn,
+    SearchFieldsFn,
+    SendFileFn,
+    UpdateProgressFn,
+} from './ai/types/aiAgentDependencies';
+import { AiAgentExploreSummary } from './ai/types/aiAgentExploreSummary';
 import {
-    isVerticalBarMetricChartConfig,
-    renderVerticalBarMetricChart,
-} from './AiService/charts/verticalBarChart';
+    getDeepLinkBlocks,
+    getExploreBlocks,
+    getFeedbackBlocks,
+    getFollowUpToolBlocks,
+} from './ai/utils/getSlackBlocks';
 import {
     AI_DEFAULT_MAX_QUERY_LIMIT,
     validateSelectedFieldsExistence,
-} from './AiService/utils/aiCopilot/validators';
+} from './ai/utils/validators';
+import { isCsvFileConfig, renderCsvFile } from './ai/visualizations/csvFile';
+import {
+    isTimeSeriesMetricChartConfig,
+    renderTimeseriesChart,
+} from './ai/visualizations/timeSeriesChart';
+import { AiAgentVizConfig } from './ai/visualizations/types';
+import {
+    isVerticalBarMetricChartConfig,
+    renderVerticalBarMetricChart,
+} from './ai/visualizations/verticalBarChart';
+import { CommercialCatalogService } from './CommercialCatalogService';
 
 type AiAgentServiceDependencies = {
+    lightdashConfig: LightdashConfig;
+    analytics: LightdashAnalytics;
+    userModel: UserModel;
     aiAgentModel: AiAgentModel;
     aiModel: AiModel;
-    slackAuthenticationModel: CommercialSlackAuthenticationModel;
     featureFlagService: FeatureFlagService;
     projectService: ProjectService;
+    catalogService: CommercialCatalogService;
     slackClient: SlackClient;
     schedulerClient: CommercialSchedulerClient;
-    analytics: LightdashAnalytics;
 };
 
 export class AiAgentService {
+    private readonly lightdashConfig: LightdashConfig;
+
+    private readonly analytics: LightdashAnalytics;
+
+    private readonly userModel: UserModel;
+
     private readonly aiAgentModel: AiAgentModel;
 
     private readonly aiModel: AiModel;
-
-    private readonly slackAuthenticationModel: CommercialSlackAuthenticationModel;
 
     private readonly featureFlagService: FeatureFlagService;
 
     private readonly projectService: ProjectService;
 
+    private readonly catalogService: CommercialCatalogService;
+
     private readonly slackClient: SlackClient;
 
     private readonly schedulerClient: CommercialSchedulerClient;
 
-    private readonly analytics: LightdashAnalytics;
-
     constructor(dependencies: AiAgentServiceDependencies) {
+        this.lightdashConfig = dependencies.lightdashConfig;
+        this.analytics = dependencies.analytics;
+        this.userModel = dependencies.userModel;
         this.aiAgentModel = dependencies.aiAgentModel;
         this.aiModel = dependencies.aiModel;
-        this.slackAuthenticationModel = dependencies.slackAuthenticationModel;
         this.featureFlagService = dependencies.featureFlagService;
+        this.catalogService = dependencies.catalogService;
         this.projectService = dependencies.projectService;
         this.slackClient = dependencies.slackClient;
         this.schedulerClient = dependencies.schedulerClient;
-        this.analytics = dependencies.analytics;
     }
 
     // from AiService getToolUtilities
@@ -451,7 +490,7 @@ export class AiAgentService {
         });
     }
 
-    async generateAgentThreadResponse(
+    async scheduleGenerateAgentThreadResponse(
         user: SessionUser,
         {
             agentUuid,
@@ -459,18 +498,18 @@ export class AiAgentService {
             prompt,
         }: {
             agentUuid: string;
-            prompt: string;
             threadUuid?: string;
+            prompt: string;
         },
     ) {
-        const { organizationUuid } = user;
-        if (!organizationUuid) {
-            throw new ForbiddenError('Organization not found');
-        }
-
         const isCopilotEnabled = await this.getIsCopilotEnabled(user);
         if (!isCopilotEnabled) {
             throw new ForbiddenError('Copilot is not enabled');
+        }
+
+        const { organizationUuid } = user;
+        if (!organizationUuid) {
+            throw new ForbiddenError('Organization not found');
         }
 
         const agent = await this.aiAgentModel.getAgent({
@@ -562,6 +601,64 @@ export class AiAgentService {
         });
 
         return { jobId, threadUuid };
+    }
+
+    async generateAgentThreadResponse({
+        agentUuid,
+        threadUuid,
+        promptUuid,
+    }: {
+        agentUuid: string;
+        threadUuid: string;
+        promptUuid: string;
+    }): Promise<void> {
+        const thread = await this.aiModel.findThread(threadUuid);
+        if (!thread) {
+            throw new NotFoundError(`Thread not found: ${threadUuid}`);
+        }
+
+        const prompt = await this.aiModel.findWebAppPrompt(promptUuid);
+        if (!prompt) {
+            throw new NotFoundError(`Prompt not found: ${promptUuid}`);
+        }
+
+        const { organizationUuid, projectUuid } = thread;
+
+        const threadMessages = await this.aiModel.getThreadMessages(
+            organizationUuid,
+            projectUuid,
+            threadUuid,
+        );
+
+        if (threadMessages.length === 0) {
+            throw new Error(
+                `No messages found in thread: ${threadUuid}. ${agentUuid} ${promptUuid}`,
+            );
+        }
+
+        const user = await this.userModel.findSessionUserAndOrgByUuid(
+            prompt.createdByUserUuid,
+            organizationUuid,
+        );
+
+        try {
+            const chatHistoryMessages =
+                getChatHistoryFromThreadMessages(threadMessages);
+
+            const response = await this.generateExploreResult(
+                user,
+                chatHistoryMessages,
+                prompt,
+            );
+
+            await this.aiModel.updateWebAppResponse({
+                promptUuid,
+                response,
+            });
+        } catch (e) {
+            Logger.error('Failed to generate agent thread response:', e);
+            throw new Error('Failed to generate agent thread response');
+        }
     }
 
     async generateAgentThreadMessageViz(
@@ -690,5 +787,674 @@ export class AiAgentService {
             promptUuid: messageUuid,
             humanScore,
         });
+    }
+
+    private async updateSlackResponseWithProgress(
+        slackPrompt: SlackPrompt,
+        progress: string,
+    ) {
+        await this.slackClient.updateMessage({
+            organizationUuid: slackPrompt.organizationUuid,
+            text: progress,
+            channelId: slackPrompt.slackChannelId,
+            messageTs: slackPrompt.response_slack_ts,
+        });
+    }
+
+    // eventually this should become a "getExplores" tool
+    private async getMinimalExploreInformation(
+        user: SessionUser,
+        projectUuid: string,
+        availableTags: string[] | null,
+    ): Promise<AiAgentExploreSummary[]> {
+        const exploreSummaries =
+            await this.projectService.getAllExploresSummary(
+                user,
+                projectUuid,
+                true,
+                false,
+            );
+
+        const explores = await Promise.all(
+            exploreSummaries.map((s) =>
+                this.projectService.getExplore(user, projectUuid, s.name),
+            ),
+        );
+
+        const exploresWithDescriptions = explores
+            .map((explore) =>
+                filterExploreByTags({
+                    explore,
+                    availableTags,
+                }),
+            )
+            .filter((explore) => explore !== undefined)
+            .map((explore, index) => ({
+                ...explore,
+                description: exploreSummaries[index]?.description,
+            }));
+
+        const minimalExploreInformation = exploresWithDescriptions.map((s) => ({
+            ...pick(s, ['name', 'label', 'description', 'baseTable']),
+            joinedTables: Object.keys(s.tables).filter(
+                (table) => table !== s.baseTable,
+            ),
+        }));
+
+        return minimalExploreInformation;
+    }
+
+    // Defines the functions that AI Agent tools can use to interact with the Lightdash backend or slack
+    // This is scoped to the project, user and prompt (closure)
+    public getAiAgentDependencies(
+        user: SessionUser,
+        prompt: SlackPrompt | AiWebAppPrompt,
+        availableTags: string[] | null,
+    ) {
+        const { projectUuid, organizationUuid, agentUuid } = prompt;
+
+        const getExplore: GetExploreFn = async ({ exploreName }) => {
+            const explore = await this.projectService.getExplore(
+                user,
+                projectUuid,
+                exploreName,
+            );
+
+            const filteredExplore = filterExploreByTags({
+                explore,
+                availableTags,
+            });
+
+            if (!filteredExplore) {
+                throw new NotFoundError('Explore not found');
+            }
+
+            return filteredExplore;
+        };
+
+        const searchFields: SearchFieldsFn = async ({
+            exploreName,
+            embeddingSearchQueries,
+        }: {
+            exploreName: string;
+            embeddingSearchQueries: Array<{
+                name: string;
+                description: string;
+            }>;
+        }) => {
+            // TODO: add to lightdash config
+            if (!process.env.OPENAI_API_KEY) {
+                throw new Error('OpenAI API key not found');
+            }
+
+            const embedQueries = await generateEmbeddingsNameAndDescription({
+                apiKey: process.env.OPENAI_API_KEY,
+                values: embeddingSearchQueries,
+            });
+
+            const catalogItems = await this.catalogService.hybridSearch({
+                user,
+                projectUuid,
+                embedQueries,
+                exploreName,
+                type: CatalogType.Field,
+                limit: 30,
+            });
+
+            return catalogItems.data.map((item) => item.name);
+        };
+
+        const updateProgress: UpdateProgressFn = (progress) =>
+            isSlackPrompt(prompt)
+                ? this.updateSlackResponseWithProgress(prompt, progress)
+                : Promise.resolve();
+
+        const getPrompt: GetPromptFn = async () => {
+            const webOrSlackPrompt = isSlackPrompt(prompt)
+                ? await this.aiModel.findSlackPrompt(prompt.promptUuid)
+                : await this.aiModel.findWebAppPrompt(prompt.promptUuid);
+
+            if (!webOrSlackPrompt) {
+                throw new Error('Prompt not found');
+            }
+            return webOrSlackPrompt;
+        };
+
+        const runMiniMetricQuery: RunMiniMetricQueryFn = async (
+            metricQuery,
+        ) => {
+            const explore = await getExplore({
+                exploreName: metricQuery.exploreName,
+            });
+
+            const metricQueryFields = [
+                ...metricQuery.dimensions,
+                ...metricQuery.metrics,
+            ];
+
+            validateSelectedFieldsExistence(explore, metricQueryFields);
+
+            return this.projectService.runMetricQuery({
+                user,
+                projectUuid,
+                metricQuery: {
+                    ...metricQuery,
+                    tableCalculations: [],
+                },
+                exploreName: metricQuery.exploreName,
+                csvLimit: metricQuery.limit,
+                context: QueryExecutionContext.AI,
+                chartUuid: undefined,
+                queryTags: {
+                    project_uuid: projectUuid,
+                    user_uuid: user.userUuid,
+                    organization_uuid: organizationUuid,
+                },
+            });
+        };
+
+        const sendFile: SendFileFn = async (args) => {
+            //
+            // TODO: https://api.slack.com/methods/files.upload does not support setting custom usernames
+            // support this in the future
+            //
+            // const agent = agentUuid
+            //     ? await this.aiAgentService.getAgent(user, agentUuid)
+            //     : undefined;
+            // let username: string | undefined;
+            // if (agent) {
+            //     username = agent.name;
+            // }
+            //
+
+            await this.slackClient.postFileToThread(args);
+        };
+
+        return {
+            getExplore,
+            searchFields: this.lightdashConfig.ai.copilot.embeddingSearchEnabled
+                ? searchFields
+                : undefined,
+            updateProgress,
+            getPrompt,
+            runMiniMetricQuery,
+            sendFile,
+        };
+    }
+
+    async generateExploreResult(
+        user: SessionUser,
+        messageHistory: CoreMessage[],
+        slackOrWebAppPrompt: SlackPrompt | AiWebAppPrompt,
+    ): Promise<string> {
+        if (!user.organizationUuid) {
+            throw new Error('Organization not found');
+        }
+
+        if (!(await this.getIsCopilotEnabled(user))) {
+            throw new Error('AI Copilot is not enabled');
+        }
+
+        // TODO: add to lightdash config
+        if (!process.env.OPENAI_API_KEY) {
+            throw new Error('OpenAI API key not found');
+        }
+
+        const { projectUuid } = slackOrWebAppPrompt;
+
+        const agentSettings =
+            'slackChannelId' in slackOrWebAppPrompt
+                ? await this.aiAgentModel.getAgentBySlackChannelId({
+                      organizationUuid: user.organizationUuid,
+                      slackChannelId: slackOrWebAppPrompt.slackChannelId,
+                  })
+                : await this.aiAgentModel.getAgent({
+                      organizationUuid: slackOrWebAppPrompt.organizationUuid,
+                      agentUuid: slackOrWebAppPrompt.agentUuid!,
+                  });
+
+        const {
+            getExplore,
+            searchFields,
+            updateProgress,
+            getPrompt,
+            runMiniMetricQuery,
+            sendFile,
+        } = this.getAiAgentDependencies(
+            user,
+            slackOrWebAppPrompt,
+            agentSettings?.tags ?? null,
+        );
+
+        const aiAgentExploreSummaries = await this.getMinimalExploreInformation(
+            user,
+            projectUuid,
+            agentSettings?.tags ?? null,
+        );
+
+        return runAgent({
+            args: {
+                openaiApiKey: process.env.OPENAI_API_KEY,
+                agentName: agentSettings.name,
+                instruction: agentSettings.instruction,
+                messageHistory,
+                aiAgentExploreSummaries,
+                promptUuid: slackOrWebAppPrompt.promptUuid,
+                maxLimit: this.lightdashConfig.query.maxLimit,
+            },
+            dependencies: {
+                getExplore,
+                searchFields,
+                runMiniMetricQuery,
+                getPrompt,
+                sendFile,
+                // avoid binding
+                updateProgress: (args) => updateProgress(args),
+                updatePrompt: (args) => this.aiModel.updateModelResponse(args),
+            },
+        });
+    }
+
+    // TODO: user permissions
+    async updateHumanScoreForPrompt(promptUuid: string, humanScore: number) {
+        await this.aiModel.updateHumanScore({
+            promptUuid,
+            humanScore,
+        });
+    }
+
+    async createSlackPrompt(data: {
+        userUuid: string;
+        projectUuid: string;
+        slackUserId: string;
+        slackChannelId: string;
+        slackThreadTs: string | undefined;
+        prompt: string;
+        promptSlackTs: string;
+        agentUuid: string | null;
+    }): Promise<[string, boolean]> {
+        let createdThread = false;
+        let threadUuid: string | undefined;
+
+        const slackPromptExists =
+            await this.aiModel.existsSlackPromptByChannelIdAndPromptTs(
+                data.slackChannelId,
+                data.promptSlackTs,
+            );
+
+        // This happens in the case of updating a prompt message in slack
+        if (slackPromptExists) {
+            throw new AiDuplicateSlackPromptError('Prompt already exists');
+        }
+
+        if (data.slackThreadTs) {
+            threadUuid =
+                await this.aiModel.findThreadUuidBySlackChannelIdAndThreadTs(
+                    data.slackChannelId,
+                    data.slackThreadTs,
+                );
+        }
+
+        if (!threadUuid) {
+            const user = await this.userModel.getUserDetailsByUuid(
+                data.userUuid,
+            );
+            if (user.organizationUuid === undefined) {
+                throw new Error('Organization not found');
+            }
+            createdThread = true;
+            threadUuid = await this.aiModel.createSlackThread({
+                organizationUuid: user.organizationUuid,
+                projectUuid: data.projectUuid,
+                createdFrom: 'slack',
+                slackUserId: data.slackUserId,
+                slackChannelId: data.slackChannelId,
+                slackThreadTs: data.slackThreadTs || data.promptSlackTs,
+                agentUuid: data.agentUuid,
+            });
+        }
+
+        if (threadUuid === undefined) {
+            throw new Error('Failed to find slack thread');
+        }
+
+        const slackTagRegex = /^.*?(<@U\d+\w*?>)/g;
+
+        const uuid = await this.aiModel.createSlackPrompt({
+            threadUuid,
+            createdByUserUuid: data.userUuid,
+            prompt: data.prompt.replaceAll(slackTagRegex, '').trim(),
+            slackUserId: data.slackUserId,
+            slackChannelId: data.slackChannelId,
+            promptSlackTs: data.promptSlackTs,
+        });
+
+        return [uuid, createdThread];
+    }
+
+    // TODO: user permissions
+    async replyToSlackPrompt(promptUuid: string): Promise<void> {
+        let slackPrompt = await this.aiModel.findSlackPrompt(promptUuid);
+        if (slackPrompt === undefined) {
+            throw new Error('Prompt not found');
+        }
+
+        await this.updateSlackResponseWithProgress(
+            slackPrompt,
+            '🤖 Thinking...',
+        );
+
+        const user = await this.userModel.findSessionUserAndOrgByUuid(
+            slackPrompt.createdByUserUuid,
+            slackPrompt.organizationUuid,
+        );
+
+        const threadMessages = await this.aiModel.getThreadMessages(
+            slackPrompt.organizationUuid,
+            slackPrompt.projectUuid,
+            slackPrompt.threadUuid,
+        );
+
+        const thread = await this.aiModel.findThread(slackPrompt.threadUuid);
+        if (!thread) {
+            throw new Error('Thread not found');
+        }
+
+        let name: string | undefined;
+        if (thread.agentUuid) {
+            const agent = await this.getAgent(user, thread.agentUuid);
+            name = agent.name;
+        }
+
+        let response: string | undefined;
+        try {
+            const chatHistoryMessages =
+                getChatHistoryFromThreadMessages(threadMessages);
+
+            response = await this.generateExploreResult(
+                user,
+                chatHistoryMessages,
+                slackPrompt,
+            );
+        } catch (e) {
+            await this.slackClient.postMessage({
+                organizationUuid: slackPrompt.organizationUuid,
+                text: `🔴 Co-pilot failed to generate a response 😥 Please try again.`,
+                channel: slackPrompt.slackChannelId,
+                thread_ts: slackPrompt.slackThreadTs,
+                username: name,
+            });
+
+            Logger.error('Failed to generate response:', e);
+            throw new Error('Failed to generate response');
+        }
+
+        if (!response) {
+            return;
+        }
+
+        // reload the slack prompt in case it was updated
+        slackPrompt = await this.aiModel.findSlackPrompt(promptUuid);
+
+        if (slackPrompt === undefined) {
+            throw new Error('Could not reload slack prompt.');
+        }
+
+        await this.slackClient.deleteMessage({
+            organizationUuid: slackPrompt.organizationUuid,
+            channelId: slackPrompt.slackChannelId,
+            messageTs: slackPrompt.response_slack_ts,
+        });
+
+        const feedbackBlocks = getFeedbackBlocks(slackPrompt);
+        const followUpToolBlocks = getFollowUpToolBlocks(slackPrompt);
+        const exploreBlocks = getExploreBlocks(
+            slackPrompt,
+            this.lightdashConfig.siteUrl,
+        );
+        const historyBlocks = getDeepLinkBlocks(
+            slackPrompt,
+            this.lightdashConfig.siteUrl,
+        );
+
+        // ! This is needed because the markdownToBlocks escapes all characters and slack just needs &, <, > to be escaped
+        // ! https://api.slack.com/reference/surfaces/formatting#escaping
+        const slackifiedMarkdown = slackifyMarkdown(response);
+
+        const newResponse = await this.slackClient.postMessage({
+            organizationUuid: slackPrompt.organizationUuid,
+            text: slackifiedMarkdown,
+            username: name,
+            channel: slackPrompt.slackChannelId,
+            thread_ts: slackPrompt.slackThreadTs,
+            unfurl_links: false,
+            blocks: [
+                {
+                    type: 'section',
+                    text: {
+                        type: 'mrkdwn',
+                        text: slackifiedMarkdown,
+                    },
+                },
+                ...exploreBlocks,
+                ...followUpToolBlocks,
+                ...feedbackBlocks,
+                ...historyBlocks,
+            ],
+        });
+
+        await this.aiModel.updateModelResponse({
+            promptUuid: slackPrompt.promptUuid,
+            response,
+        });
+
+        if (newResponse.ts) {
+            await this.aiModel.updateSlackResponseTs({
+                promptUuid: slackPrompt.promptUuid,
+                responseSlackTs: newResponse.ts,
+            });
+        }
+    }
+
+    async getConversations(
+        user: SessionUser,
+        projectUuid: string,
+    ): Promise<AiConversation[]> {
+        if (!(await this.getIsCopilotEnabled(user))) {
+            throw new Error('AI Copilot is not enabled');
+        }
+
+        if (!user.organizationUuid) {
+            throw new Error('Organization not found');
+        }
+
+        // TODO: this is a temporary solution to check project permissions...
+        const projectSummary = await this.projectService.getProject(
+            projectUuid,
+            user,
+        );
+
+        const threads = await this.aiModel.getThreads(
+            user.organizationUuid,
+            projectUuid,
+        );
+
+        return threads.map((thread) => ({
+            threadUuid: thread.ai_thread_uuid,
+            createdAt: thread.created_at,
+            createdFrom: thread.created_from,
+            firstMessage: thread.prompt,
+            user: {
+                uuid: thread.user_uuid,
+                name: thread.user_name,
+            },
+        }));
+    }
+
+    async getConversationMessages(
+        user: SessionUser,
+        projectUuid: string,
+        aiThreadUuid: string,
+    ): Promise<AiConversationMessage[]> {
+        if (!(await this.getIsCopilotEnabled(user))) {
+            throw new Error('AI Copilot is not enabled');
+        }
+
+        const { organizationUuid } = user;
+
+        if (!organizationUuid) {
+            throw new Error('Organization not found');
+        }
+
+        const canViewProject = user.ability.can(
+            'view',
+            subject('Project', {
+                organizationUuid,
+                projectUuid,
+            }),
+        );
+
+        if (!canViewProject) {
+            throw new Error('User does not have access to the project!');
+        }
+
+        const messages = await this.aiModel.getThreadMessages(
+            organizationUuid,
+            projectUuid,
+            aiThreadUuid,
+        );
+
+        return messages.map((message) => ({
+            promptUuid: message.ai_prompt_uuid,
+            message: message.prompt,
+            createdAt: message.created_at,
+            response: message.response ?? undefined,
+            respondedAt: message.responded_at ?? undefined,
+            vizConfigOutput: message.viz_config_output ?? undefined,
+            filtersOutput: message.filters_output ?? undefined,
+            metricQuery: message.metric_query ?? undefined,
+            humanScore: message.human_score ?? undefined,
+            user: {
+                uuid: message.user_uuid,
+                name: message.user_name,
+            },
+        }));
+    }
+
+    private async _createWebAppPrompt(
+        user: LightdashUser,
+        projectUuid: string,
+        prompt: string,
+    ): Promise<string> {
+        if (!user.organizationUuid) {
+            throw new Error('Organization not found');
+        }
+
+        const threadUuid = await this.aiModel.createWebAppThread({
+            organizationUuid: user.organizationUuid,
+            projectUuid,
+            userUuid: user.userUuid,
+            createdFrom: 'web_app',
+            agentUuid: null,
+        });
+
+        if (threadUuid === undefined) {
+            throw new Error('Failed to create web app thread');
+        }
+
+        const promptUuid = await this.aiModel.createWebAppPrompt({
+            threadUuid,
+            createdByUserUuid: user.userUuid,
+            prompt,
+        });
+
+        return promptUuid;
+    }
+
+    async createWebAppConversation(
+        user: SessionUser,
+        projectUuid: string,
+        question: string,
+    ): Promise<{
+        prompt: AiWebAppPrompt;
+        rows: Record<string, AnyType>[] | undefined;
+    }> {
+        if (!(await this.getIsCopilotEnabled(user))) {
+            throw new Error('AI Copilot is not enabled');
+        }
+
+        const userDetails = await this.userModel.getUserDetailsByUuid(
+            user.userUuid,
+        );
+
+        if (userDetails.organizationUuid === undefined) {
+            throw new Error('Organization not found');
+        }
+
+        // TODO: relax this permission after we have a proper UI for copilot/project access
+        const canManageProject = user.ability.can(
+            'manage',
+            subject('Project', {
+                organizationUuid: userDetails.organizationUuid,
+                projectUuid,
+            }),
+        );
+
+        if (!canManageProject) {
+            throw new Error('User does not have access to the project!');
+        }
+
+        await this.aiModel.createWebAppThread({
+            organizationUuid: userDetails.organizationUuid,
+            projectUuid,
+            createdFrom: 'web_app',
+            userUuid: user.userUuid,
+            agentUuid: null,
+        });
+
+        const promptUuid = await this._createWebAppPrompt(
+            user,
+            projectUuid,
+            question,
+        );
+
+        const webAppPrompt = await this.aiModel.findWebAppPrompt(promptUuid);
+
+        if (!webAppPrompt) {
+            throw new Error('Prompt not found');
+        }
+
+        const response = await this.generateExploreResult(
+            user,
+            [{ role: 'user', content: question }],
+            webAppPrompt,
+        );
+
+        await this.aiModel.updateWebAppResponse({
+            promptUuid: webAppPrompt.promptUuid,
+            response,
+        });
+
+        const finalPrompt = await this.aiModel.findWebAppPrompt(promptUuid);
+
+        if (!finalPrompt) {
+            throw new Error('Prompt not found');
+        }
+
+        // TODO: this can receive project settings/available tags - like we have for slack
+        const utils = this.getAiAgentDependencies(user, finalPrompt, null);
+
+        const rows = finalPrompt.metricQuery
+            ? (
+                  await utils.runMiniMetricQuery(
+                      finalPrompt.metricQuery as AiMetricQuery,
+                  )
+              ).rows
+            : undefined;
+
+        return {
+            prompt: finalPrompt,
+            rows,
+        };
     }
 }
