+++ packages/backend/src/ee/services/ai/models/openai-gpt-4.1.ts
@@ -1,0 +1,14 @@
+import { createOpenAI } from '@ai-sdk/openai';
+
+export const getOpenaiGpt41model = (apiKey: string) => {
+    const openai = createOpenAI({
+        apiKey,
+        compatibility: 'strict',
+    });
+
+    const model = openai('gpt-4.1', {
+        structuredOutputs: true,
+    });
+
+    return model;
+};
