+++ packages/backend/src/ee/services/ai/types/followUpTools.ts
