+++ packages/backend/src/ee/services/ai/visualizations/renderEcharts.ts
